Phoenix Shield item base
