+++ src/app/core/items/neutral/phoenix-shield.ts
@@ -1,0 +1,13 @@
+import { ItemSlotType } from '../types';
+import { createItem } from '../utils';
+
+
+export const PhoenixShieldItem = createItem({
+  name: 'Phoenix Shield',
+  icon: 'fire-shield',
+  slot: ItemSlotType.Shield,
+  stats: {
+    playerBonusDefence: 2,
+    playerBonusAttack: 1,
+  },
+});
