+++ src/app/core/structures/common/index.ts
@@ -6,4 +6,4 @@ export * from './graveyard';
 export * from './magic-river';
 export * from './mountain-nest';
 export * from './witch-hut';
-
+export * from './thiefs-lair';
